pyproject.toml
# Python project metadata, including dependencies & tool settings
# https://snarky.ca/what-the-heck-is-pyproject-toml

[project]
name = "blub"
version = "0.1"
requires-python = ">=3.12"

# Dependencies needed at runtime
dependencies = [
  "ok-logging-setup>=0.2",
  "paho-mqtt~=2.1",
  "piexif~=1.1",
  "pint~=0.24",
  "pyserial~=3.5",
  "python-debian>=1.0.1",
]
